pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Analyze web pages"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
]

[tool.ruff]
line-length = 80
indent-width = 4
